// src/services/CoingeckoService/CoingeckoService.types.ts
export type CoingeckoCoin = {
  id: string;
  name: string;
  symbol: string;
  image: string;
  current_price: number;
  market_cap: number;
  market_cap_rank: number;
  total_volume: number;
  high_24h: number;
  low_24h: number;
  price_change_24h: number;
  price_change_percentage_24h: number;
  market_cap_change_24h: number;
  market_cap_change_percentage_24h: number;
  circulating_supply: number;
  total_supply: number;
  ath: number;
  ath_change_percentage: number;
  ath_date: string;
  roi: {
    times: number;
    currency: string;
  };
  last_updated: string;
};

export type Coin = {
  id: string;
  name: string;
  symbol: string;
  price: number;
  marketCap: number;
  imageUrl: string;
};